+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Nothing yet
+
+## [0.8.0] - 2023-04-10
+
 ### Changed
 
 - PeriodType renamed Period
@@ -67,7 +71,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests with well-known MPDs
 - Tweaked XML library to support namespaces
 
-[Unreleased]: https://github.com/Eyevinn/dash-mpd/compare/v0.7.0...HEAD
+[Unreleased]: https://github.com/Eyevinn/dash-mpd/compare/v0.8.0...HEAD
+[0.8.0]: https://github.com/Eyevinn/dash-mpd/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/Eyevinn/dash-mpd/compare/v0.6.1...v0.7.0
 [0.6.1]: https://github.com/Eyevinn/dash-mpd/compare/v0.6.0...v0.6.1
 [0.6.0]: https://github.com/Eyevinn/dash-mpd/compare/v0.5.0...v0.6.0